gl-430/texture-view - Vertex shader error precision qualifier can't be use in constructor.
Precision qualifier only can be in
- Variable declaration
- Function return type
- Function parameter

GLSL spec 4.4, chapeter 4.7, section 2, page 83 said:

"Precision qualifiers, as with other qualifiers, do not effect the basic
type of the variable. In particular, there are no constructors for
precision conversions; constructors only convert types. Similarly,
precision qualifiers, as with other qualifiers, do not contribute to
function overloading based on parameter types. As discussed in the next
chapter, function input and output is done through copies, and therefore
qualifiers do not have to match."