Update buffer-storage.vert
Shared qualifier is the default one.  Shared behave the same way as packed, which give the freedom to the driver to rearrange, compact or otherwise mess with the layout of the buffer content, with two condition the process it’s consistent, and doesn't remove any members. The application is assuming that std430 is the default one.